Make namespace HydroSmallVel
