+++ EBGodunov/hydro_ebgodunov_extrap_vel_to_faces.cpp
@@ -15,7 +15,9 @@
 #include <hydro_godunov.H>
 #include <hydro_godunov_K.H>
 #include <hydro_bcs_K.H>
+#include <hydro_small_vel.H>
 
+using namespace HydroSmallVel;
 using namespace amrex;
 void
 EBGodunov::ExtrapVelToFaces ( MultiFab const& vel,
