+++ Godunov/hydro_godunov_edge_state_3D.cpp
@@ -11,7 +11,9 @@
 #include <hydro_godunov_corner_couple.H>
 #include <hydro_godunov_K.H>
 #include <hydro_bcs_K.H>
+#include <hydro_small_vel.H>
 
+using namespace HydroSmallVel;
 using namespace amrex;
 
 void
